Fix line/(poly?) intersections
These should work perfectly now

+++ src/alfredo/geom/Point.java
@@ -1,11 +1,11 @@
 package alfredo.geom;
 
 public class Point {
-    public int x, y;
+    public float x, y;
     
     public Point() { x = 0; y = 0; }
     
-    public Point(int x, int y) {
+    public Point(float x, float y) {
         this.x = x;
         this.y = y;
     }
